requirements.txt
discord.py[voice]==1.5.1 # discord
async-timeout==3.0.1 # discord
pytz==2020.4 # timezone
praw==7.1.0 # reddit
youtube-dl==2020.11.17 # music
git+https://github.com/johnwmillr/LyricsGenius.git # lyrics
